+++ pkg/app/tests/ds/ds_test.go
@@ -2,22 +2,17 @@ package ds_test
 
 import (
 	"context"
-	"path/filepath"
 	"testing"
 	"time"
+
+	azc "github.com/aserto-dev/topaz/pkg/cli/clients/authorizer"
+	dsc "github.com/aserto-dev/topaz/pkg/cli/clients/directory"
 
 	client "github.com/aserto-dev/go-aserto"
 	dsr3 "github.com/aserto-dev/go-directory/aserto/directory/reader/v3"
 	assets_test "github.com/aserto-dev/topaz/pkg/app/tests/assets"
 	tc "github.com/aserto-dev/topaz/pkg/app/tests/common"
-	"github.com/aserto-dev/topaz/pkg/cli/cc"
-	dsc "github.com/aserto-dev/topaz/pkg/cli/clients/directory"
-	"github.com/aserto-dev/topaz/pkg/cli/cmd/common"
-	"github.com/aserto-dev/topaz/pkg/cli/cmd/directory"
-	"github.com/aserto-dev/topaz/pkg/cli/cmd/templates"
-	"google.golang.org/protobuf/types/known/structpb"
 
-	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"github.com/testcontainers/testcontainers-go"
 	"github.com/testcontainers/testcontainers-go/wait"
@@ -64,16 +59,30 @@ func TestDirectory(t *testing.T) {
 		cancel()
 	})
 
-	grpcAddr, err := tc.MappedAddr(ctx, topaz, "9292")
+	addr, err := tc.MappedAddr(ctx, topaz, "9292")
 	require.NoError(t, err)
 
-	t.Run("testDirectory", testDirectory(grpcAddr))
+	dsConfig := &dsc.Config{
+		Host:      addr,
+		Insecure:  true,
+		Plaintext: false,
+		Timeout:   10 * time.Second,
+	}
+
+	azConfig := &azc.Config{
+		Host:      addr,
+		Insecure:  true,
+		Plaintext: false,
+		Timeout:   10 * time.Second,
+	}
+
+	t.Run("testDirectory", testDirectory(dsConfig, azConfig))
 }
 
-func testDirectory(addr string) func(*testing.T) {
+func testDirectory(dsConfig *dsc.Config, azConfig *azc.Config) func(*testing.T) {
 	return func(t *testing.T) {
 		opts := []client.ConnectionOption{
-			client.WithAddr(addr),
+			client.WithAddr(dsConfig.Host),
 			client.WithInsecure(true),
 		}
 
@@ -84,7 +93,7 @@ func testDirectory(addr string) func(*testing.T) {
 		ctx, cancel := context.WithCancel(context.Background())
 		t.Cleanup(cancel)
 
-		t.Run("", installTemplate(addr, "../../../../assets/gdrive.json"))
+		t.Run("", tc.InstallTemplate(dsConfig, azConfig, "../../../../assets/gdrive.json"))
 
 		tests := []struct {
 			name string
@@ -97,101 +106,5 @@ func testDirectory(addr string) func(*testing.T) {
 		for _, testCase := range tests {
 			t.Run(testCase.name, testCase.test)
 		}
-	}
-}
-
-func installTemplate(addr, tmpl string) func(*testing.T) {
-	return func(t *testing.T) {
-		t.Logf("addr: %s tmpl: %s", addr, tmpl)
-
-		t.Setenv("TOPAZ_NO_COLOR", "true")
-		c, err := cc.NewCommonContext(context.Background(), true, filepath.Join(cc.GetTopazDir(), common.CLIConfigurationFile))
-		require.NoError(t, err)
-
-		dsConfig := &dsc.Config{
-			Host:      addr,
-			Insecure:  true,
-			Plaintext: false,
-			Timeout:   10 * time.Second,
-		}
-
-		absPath, err := filepath.Abs(tmpl)
-		require.NoError(t, err)
-
-		tmpl, err := templates.GetTemplateFromFile(absPath)
-		require.NoError(t, err)
-
-		t.Logf("name %s", tmpl.Name)
-		t.Logf("template: %s", absPath)
-
-		dirPath := filepath.Dir(absPath)
-		t.Logf("dir %s", dirPath)
-
-		manifestFile := filepath.Join(dirPath, tmpl.Assets.Manifest)
-		t.Logf("manifestFile: %s", manifestFile)
-		t.Run(tmpl.Name+"-DeleteManifest", DeleteManifest(c, dsConfig))
-		t.Run(tmpl.Name+"-SetManifest", SetManifest(c, dsConfig, manifestFile))
-
-		if len(tmpl.Assets.IdentityData) > 0 {
-			idpDataDir := filepath.Dir(filepath.Join(dirPath, tmpl.Assets.IdentityData[0]))
-			t.Logf("idp_data: %s", idpDataDir)
-			t.Run(tmpl.Name+"-ImportIdentityData", ImportData(c, dsConfig, idpDataDir))
-		}
-
-		if len(tmpl.Assets.DomainData) > 0 {
-			domainDataDir := filepath.Dir(filepath.Join(dirPath, tmpl.Assets.DomainData[0]))
-			t.Logf("domain_data: %s", domainDataDir)
-			t.Run(tmpl.Name+"-ImportDomainData", ImportData(c, dsConfig, domainDataDir))
-		}
-
-		if len(tmpl.Assets.Assertions) > 0 {
-			assertionsFile := filepath.Join(dirPath, tmpl.Assets.Assertions[0])
-			t.Logf("assertionsFile: %s", assertionsFile)
-			t.Run(tmpl.Name+"-ExecDirectoryTest", ExecDirectoryTests(c, dsConfig, []string{assertionsFile}))
-		}
-	}
-}
-
-func DeleteManifest(c *cc.CommonCtx, cfg *dsc.Config) func(*testing.T) {
-	return func(t *testing.T) {
-		cmd := directory.DeleteManifestCmd{Config: *cfg, Force: true}
-		if err := cmd.Run(c); err != nil {
-			assert.NoError(t, err)
-		}
-	}
-}
-
-func SetManifest(c *cc.CommonCtx, cfg *dsc.Config, path string) func(*testing.T) {
-	return func(t *testing.T) {
-		cmd := directory.SetManifestCmd{Config: *cfg, Path: path}
-		if err := cmd.Run(c); err != nil {
-			assert.NoError(t, err)
-		}
-	}
-}
-
-func ImportData(c *cc.CommonCtx, cfg *dsc.Config, dir string) func(*testing.T) {
-	return func(t *testing.T) {
-		cmd := directory.ImportCmd{Config: *cfg, Directory: dir}
-		if err := cmd.Run(c); err != nil {
-			assert.NoError(t, err)
-		}
-	}
-}
-
-func ExecDirectoryTests(c *cc.CommonCtx, cfg *dsc.Config, files []string) func(*testing.T) {
-	return func(t *testing.T) {
-		cmd := directory.TestExecCmd{Config: *cfg, TestExecCmd: common.TestExecCmd{Files: files, Summary: true, Desc: "on-error"}}
-		if err := cmd.Run(c); err != nil {
-			assert.NoError(t, err)
-		}
-	}
-}
-
-func SetContext(k, v string) *structpb.Struct {
-	return &structpb.Struct{
-		Fields: map[string]*structpb.Value{
-			k: structpb.NewStringValue(v),
-		},
 	}
 }
